Make Message and Client structs public.
This is necessary to be able to do things like the following:

 ch := make(chan *irc.Client)
 // do some other work, obtain a *Client object named client
 ch<- client

Previously this wouldn't work because irc.Client was an interface and client was
a private struct in the irc package. A pointer to irc.Client is not the same
thing as a pointer to irc.client so the compiler was complaining loudly.

I left the old Client interface commented out rather than deleting it in case
there is a use for it as an interface.

Similar adjustments were made to irc.Message/irc.message

Signed-off-by: Wayne Warren <waynr@sdf.org>